Move the parsing to a different file.
TODO<joka921> parallelization and testing.
