+++ src/spatialjoin/SpatialJoinMain.cpp
@@ -3,18 +3,14 @@
 // Authors: Patrick Brosi <brosi@informatik.uni-freiburg.de>
 
 #include <iostream>
-#include <string_view>
 #include <optional>
 
-#include <absl/strings/charconv.h>
-
-#include "BoxIds.h"
 #include "Sweeper.h"
 #include "util/Misc.h"
 #include "util/geo/Geo.h"
 #include "util/http/Server.h"
 #include "util/log/Log.h"
-#include "ctre.hpp"
+#include "WktParser.h"
 
 using sj::Sweeper;
 using util::geo::DLine;
@@ -28,74 +24,6 @@ static const char *YEAR = &__DATE__[7];
 static const char *COPY =
         "University of Freiburg - Chair of Algorithms and Data Structures";
 static const char *AUTHORS = "Patrick Brosi <brosi@informatik.uni-freiburg.de>";
-
-constexpr auto fs(const auto &s) {
-    return ctll::fixed_string{s};
-}
-
-/// Helper function for ctre: concatenation of fixed_strings
-template<size_t A, size_t B>
-constexpr ctll::fixed_string<A + B> operator+(const ctll::fixed_string<A> &a,
-                                              const ctll::fixed_string<B> &b) {
-    char32_t comb[A + B + 1] = {};
-    for (size_t i = 0; i < A; ++i) {  // omit the trailing 0 of the first string
-        comb[i] = a.content[i];
-    }
-    for (size_t i = 0; i < B; ++i) {
-        comb[i + A] = b.content[i];
-    }
-    // the input array must be zero-terminated
-    comb[A + B] = '\0';
-    return ctll::fixed_string(comb);
-}
-
-/// Helper function for ctre: concatenation of fixed_strings
-template<size_t A, size_t B>
-constexpr ctll::fixed_string<A + B - 1> operator+(
-        const ctll::fixed_string<A> &a, const char (&b)[B]) {
-    auto bStr = ctll::fixed_string(b);
-    return a + bStr;
-}
-
-/// Helper function for ctre: concatenation of fixed_strings
-template<size_t A, size_t B>
-constexpr ctll::fixed_string<A + B - 1> operator+(
-        const char (&a)[A], const ctll::fixed_string<B> &b) {
-    auto aStr = ctll::fixed_string(a);
-    return aStr + b;
-}
-
-/// Create a regex group by putting the argument in parentheses
-template<size_t N>
-static constexpr auto grp(const ctll::fixed_string<N> &s) {
-    return ctll::fixed_string("(") + s + ctll::fixed_string(")");
-}
-
-template<size_t N, size_t M>
-static constexpr auto grp(const ctll::fixed_string<N> &s, const ctll::fixed_string<M> &groupName) {
-    return grp(ctll::fixed_string("?<") + groupName + ctll::fixed_string(">") + s);
-}
-
-/// Create a regex character class by putting the argument in square brackets
-template<size_t N>
-static constexpr auto cls(const ctll::fixed_string<N> &s) {
-    return ctll::fixed_string("[") + s + ctll::fixed_string("]");
-}
-
-template<size_t N>
-static constexpr auto paren(const ctll::fixed_string<N> &s) {
-    return ctll::fixed_string("\\(") + s + ctll::fixed_string("\\)");
-}
-
-template<size_t N>
-static constexpr auto optWsAround(const ctll::fixed_string<N> &s) {
-    return ctll::fixed_string("\\s*") + s + ctll::fixed_string("\\s*");
-}
-
-template<size_t N>
-static constexpr auto star(const ctll::fixed_string<N> &s) {
-    return s + fs("*");
-}
 
 // _____________________________________________________________________________
 void printHelp(int argc, char **argv) {
@@ -131,361 +59,9 @@ void printHelp(int argc, char **argv) {
 }
 
 // _____________________________________________________________________________
-util::geo::I32Line parseLineString(const std::string &a, size_t p) {
-    util::geo::I32Line line;
-    line.reserve(2);
-    auto end = memchr(a.c_str() + p, ')', a.size() - p);
-    assert(end);
-
-    while (true) {
-        while (*(a.c_str() + p) && isspace(*(a.c_str() + p))) p++;
-        double x = util::atof(a.c_str() + p, 10);
-        double y = util::atof(
-                static_cast<const char *>(memchr(a.c_str() + p, ' ', a.size() - p)) + 1,
-                10);
-        auto projPoint = latLngToWebMerc(DPoint(x, y));
-
-        line.push_back(I32Point{projPoint.getX() * PREC, projPoint.getY() * PREC});
-
-        auto n = memchr(a.c_str() + p, ',', a.size() - p);
-        if (!n || n > end) break;
-        p = static_cast<const char *>(n) - a.c_str() + 1;
-    }
-
-    return util::geo::simplify(line, 0);
-}
-
-// _____________________________________________________________________________
-util::geo::I32Point parsePoint(const std::string &a, size_t p) {
-    auto point = latLngToWebMerc(DPoint(
-            util::atof(a.c_str() + p, 10),
-            util::atof(
-                    static_cast<const char *>(memchr(a.c_str() + p, ' ', a.size() - p)) +
-                    1,
-                    10)));
-
-    return {point.getX() * PREC, point.getY() * PREC};
-}
-
-static constexpr ctll::fixed_string floatRegex = "[0-9]+(\\.[0-9]+)?";
-static constexpr ctll::fixed_string coordinateRegex =
-        grp(floatRegex, fs("n1")) + fs("\\s+") + grp(floatRegex, fs("n2"));
-static constexpr ctll::fixed_string innerParens = paren(star(cls(fs("^)"))));
-
-std::optional<I32Point> matchPoint(std::string_view input) {
-    static constexpr ctll::fixed_string pointRegex = fs("\\s*POINT\\s*\\(") + coordinateRegex + fs("\\)");
-    static_assert(ctre::match<pointRegex>("POINT (3.102 30)"));
-    auto match = ctre::match<pointRegex>(input);
-    if (!match) {
-        return std::nullopt;
-    }
-    float f1;
-    float f2;
-    const auto &m1 = match.get<"n1">().to_view();
-    const auto &m2 = match.get<"n2">().to_view();
-    absl::from_chars(m1.data(), m1.data() + m1.size(), f1);
-    absl::from_chars(m2.data(), m2.data() + m2.size(), f2);
-    auto point = latLngToWebMerc(DPoint(
-            f1, f2));
-    return I32Point{point.getX() * PREC, point.getY() * PREC};
-}
-
-I32Line parseCoordinateList(std::string_view input) {
-    I32Line line;
-    for (const auto &m: ctre::range<coordinateRegex>(input)) {
-        float f1;
-        float f2;
-        const auto &m1 = m.get<"n1">().to_view();
-        const auto &m2 = m.get<"n2">().to_view();
-        absl::from_chars(m1.data(), m1.data() + m1.size(), f1);
-        absl::from_chars(m2.data(), m2.data() + m2.size(), f2);
-        auto projPoint = latLngToWebMerc(DPoint(f1, f2));
-        line.push_back(I32Point{projPoint.getX() * PREC, projPoint.getY() * PREC});
-    }
-    return line;
-}
-
-std::optional<I32Line> matchLinestring(std::string_view input) {
-    static_assert(ctre::starts_with<"\\s*LINESTRING">("LINESTRING(.asdfj)"));
-    if (!ctre::starts_with<"\\s*LINESTRING">(input)) {
-        return std::nullopt;
-    }
-
-    // TODO<joka921> We could also remove the beginning first
-    return parseCoordinateList(input);
-}
-
-I32Polygon matchPolygonInner(std::string_view input) {
-    std::vector<I32Line> lines;
-    size_t position = 0;
-    while (true) {
-        position = input.find('(', position + 1);
-        if (position == std::string_view::npos) {
-            break;
-        }
-        while (true) {
-            auto close = input.find_first_of("()", position + 1);
-            if (close == std::string_view::npos) {
-                throw std::runtime_error("unmatched parentheses");
-            }
-            if (input[close] == '(') {
-                position = close;
-                continue;
-            }
-            lines.push_back(parseCoordinateList(std::string_view{input.data() + position, input.data() + close}));
-            break;
-        }
-    }
-    /*
-    for (const auto &m: ctre::range<"\\([^(]*\\)">(input)) {
-        lines.push_back(parseCoordinateList(m.to_view()));
-    }
-     */
-    I32Polygon result;
-    result.getOuter() = std::move(lines.at(0));
-    result.getInners().reserve(lines.size() - 1);
-    std::move(lines.begin() + 1, lines.end(), std::back_inserter(result.getInners()));
-    return result;
-}
-
-std::optional<I32Polygon> matchPolygon(std::string_view input) {
-    if (!ctre::starts_with<"\\s*POLYGON">(input)) {
-        return std::nullopt;
-    }
-
-    return matchPolygonInner(input);
-}
-
-std::optional<I32MultiPolygon> matchMultipolygon(std::string_view input) {
-    /*
-    if (input.size() > 200'000) {
-        return std::nullopt;
-    }
-     */
-
-    if (const auto &m = ctre::starts_with<"\\s*MULTIPOLYGON\\s*\\(">(input); !m) {
-        return std::nullopt;
-    } else {
-        input.remove_prefix(m.to_view().size());
-    }
-
-    static constexpr auto polyStructure = paren(innerParens + grp(optWsAround(fs(",")) + innerParens) + fs("*"));
-
-    static_assert(ctre::match<innerParens>("( asdif )"));
-    static_assert(ctre::match<"[^(]">("y"));
-
-    std::vector<I32Polygon> polygons;
-    for (const auto &m: ctre::range<polyStructure>(input)) {
-        polygons.push_back(matchPolygonInner(m.to_view()));
-    }
-    return I32MultiPolygon{std::move(polygons)};
-}
-
-// _____________________________________________________________________________
-void parse(const char *c, size_t size, std::string &dangling, size_t *gid,
-           Sweeper &idx) {
-    const char *start = c;
-    while (c < start + size) {
-        if (*c == '\n') {
-            (*gid)++;
-
-            auto idp = dangling.find("\t");
-
-            std::string id = std::to_string(*gid);
-
-            size_t start = 2;
-
-            if (idp != std::string::npos) {
-                id = dangling.substr(0, idp);
-                start = idp + 2;
-            }
-
-            if (auto p = matchPoint(dangling)) {
-
-            }
-            auto p = dangling.rfind("POINT(", start);
-
-            if (p != std::string::npos) {
-                p += 6;
-                auto point = parsePoint(dangling, p);
-                idx.add(point, id);
-            } else if ((p = dangling.rfind("LINESTRING(", start)) !=
-
-                       std::string::npos) {
-                p += 11;
-                const auto &line = parseLineString(dangling, p);
-                if (line.size() != 0) {
-                    idx.add(line, id);
-                }
-            } else if ((p = dangling.rfind("MULTILINESTRING(", start)) !=
-                       std::string::npos) {
-                p += 16;
-                size_t i = 0;
-                while ((p = dangling.find("(", p + 1)) != std::string::npos) {
-                    const auto &line = parseLineString(dangling, p + 1);
-                    if (line.size() != 0) {
-                        // TODO, i is the line number
-                    }
-                    i++;
-                }
-            } else if ((p = dangling.rfind("POLYGON(", start)) != std::string::npos) {
-                p += 7;
-                size_t i = 0;
-                I32Polygon poly;
-                while ((p = dangling.find("(", p + 1)) != std::string::npos) {
-                    const auto &line = parseLineString(dangling, p + 1);
-                    if (i == 0) {
-                        // outer
-                        poly.getOuter() = line;
-                    } else {
-                        poly.getInners().push_back(line);
-                    }
-                    i++;
-                }
-                idx.add(poly, id);
-            } else if ((p = dangling.rfind("MULTIPOLYGON(", start)) !=
-                       std::string::npos) {
-                p += 12;
-                I32MultiPolygon mp;
-                while (p != std::string::npos &&
-                       (p = dangling.find("(", p + 1)) != std::string::npos) {
-                    I32Polygon poly;
-                    size_t i = 0;
-                    while ((p = dangling.find("(", p + 1)) != std::string::npos) {
-                        const auto &line = parseLineString(dangling, p + 1);
-                        if (i == 0) {
-                            // outer
-                            poly.getOuter() = line;
-                        } else {
-                            poly.getInners().push_back(line);
-                        }
-
-                        // check if multipolygon is closed
-                        auto q = dangling.find(
-                                ")", p + 1);  // this is the closing of the linestring
-                        auto q2 = dangling.find(")", q + 1);
-                        auto q3 = dangling.find(",", q + 1);
-                        if (q2 != std::string::npos && q3 != std::string::npos && q2 < q3) {
-                            p = q3;
-                            break;
-                        }
-
-                        i++;
-                    }
-                    mp.push_back(poly);
-                }
-                idx.add(mp, id);
-            }
-
-            dangling.clear();
-            c++;
-            continue;
-        }
-
-        dangling += *c;
-
-        c++;
-    }
-}
-
-// _____________________________________________________________________________
-void parseElement(std::string_view line, std::string gid,
-                  Sweeper &idx) {
-    if (auto p = matchPoint(line)) {
-        idx.add(p.value(), std::move(gid));
-        return;
-    } else if (auto l = matchLinestring(line)) {
-        idx.add(l.value(), std::move(gid));
-    } else if (auto poly = matchPolygon(line)) {
-        idx.add(poly.value(), std::move(gid));
-    } else if (auto mpoly = matchMultipolygon(line)) {
-        idx.add(mpoly.value(), std::move(gid));
-    } else {
-        std::cerr << "Couldn't parse line of size " << line.size() << std::endl;
-        std::cerr << "Couldn't parse element \""  << line.substr(0, 40) << std::endl;
-        return;
-    }
-}
-
-// _____________________________________________________________________________
-void parseNew(std::string_view &input, size_t *gid,
-              Sweeper &idx) {
-    while (true) {
-        auto idp = input.find('\t');
-        auto newline = input.find('\n');
-
-        if (newline == std::string_view::npos) {
-            return;
-            // TODO<joka921> Fix the rest.
-        }
-
-        auto line = input.substr(0, newline);
-        input.remove_prefix(newline + 1);
-
-
-        std::string id;
-        if (idp != std::string_view::npos) {
-            id = line.substr(0, idp);
-            line.remove_prefix(idp + 1);
-        } else {
-            id = std::to_string(*gid);
-        }
-
-        parseElement(line, std::move(id), idx);
-        ++(*gid);
-    }
-}
-
-void parseStdin(Sweeper &idx) {
-    std::vector<char> buffer;
-    buffer.resize(100'000'000);
-    size_t offset = 0;
-    // The lines by default were 1-based...
-    size_t gid = 1;
-    while (true) {
-        size_t toRead = buffer.size() - offset;
-        size_t actualOffset = offset;
-        while (true) {
-            auto numRead = read(0, buffer.data() + actualOffset, toRead);
-            toRead -= numRead;
-            actualOffset += numRead;
-            if (toRead == 0) {
-                // read a complete block, break;
-                break;
-            }
-            //std::cerr << "to Read" << toRead << "num Read " << numRead << std::endl;
-            if (numRead == 0) {
-                if (offset + numRead == 0) {
-                    return;
-                }
-                if (buffer.at(offset + numRead - 1) != '\n') {
-                    buffer.push_back('\n');
-                    ++numRead;
-                }
-                std::string_view sv{buffer.data(), offset + numRead};
-                //std::cerr << "parsing " << sv.size() << "bytes " << std::endl;
-                parseNew(sv, &gid, idx);
-                return;
-            }
-        }
-        std::string_view sv{buffer.data(), buffer.size()};
-        //std::cerr << "parsing " << sv.size() << "bytes(intermediate)" << std::endl;
-        parseNew(sv, &gid, idx);
-        auto beg = sv.data() - buffer.data();
-        auto end = beg + sv.size();
-        offset = sv.size();
-        std::shift_left(buffer.data(), buffer.data() + end, beg);
-    }
-}
-
-
-// _____________________________________________________________________________
 int main(int argc, char **argv) {
     // disable output buffering for standard output
     setbuf(stdout, NULL);
-
-    auto p = matchMultipolygon(
-            "MULTIPOLYGON(((7.4636964 51.4194415,7.4636994 51.4195635,7.4637154 51.4200254,7.4637234 51.4201014,7.4637354 51.4202044,7.4637544 51.4203684,7.4637714 51.4205364,7.4638574 51.4214862,7.4638564 51.4218412,7.4638034 51.4230121,7.4637744 51.4234130,7.4637754 51.4235350,7.4637524 51.4236910,7.4638464 51.4246329,7.4643244 51.4254730,7.4649252 51.4259337,7.4649262 51.4268346,7.4648324 51.4270101,7.4648346 51.4271056,7.4648074 51.4274032,7.4647949 51.4278262,7.4647242 51.4280135,7.4648892 51.4284635,7.4649362 51.4285934,7.4650442 51.4287874,7.4650922 51.4288754,7.4652250 51.4290814,7.4652900 51.4291789,7.4653590 51.4292730,7.4655700 51.4295702,7.4656505 51.4296540,7.4657925 51.4298390,7.4658990 51.4299201,7.4661393 51.4301292,7.4664204 51.4303978,7.4666129 51.4305922,7.4669459 51.4308662,7.4671928 51.4310692,7.4674678 51.4312941,7.4674578 51.4313811,7.4674338 51.4315841,7.4673388 51.4319841,7.4672688 51.4322860,7.4673518 51.4323120,7.4674028 51.4324356,7.4674338 51.4324840,7.4674258 51.4325870,7.4669298 51.4344157,7.4668067 51.4345165,7.4670088 51.4345600,7.4669864 51.4345379,7.4672608 51.4345964,7.4678994 51.4347327,7.4684172 51.4348432,7.4688713 51.4349401,7.4690722 51.4349895,7.4692083 51.4350205,7.4692835 51.4350376,7.4694783 51.4350951,7.4696804 51.4351596,7.4696978 51.4351652,7.4698519 51.4352142,7.4702339 51.4353360,7.4706588 51.4354716,7.4711792 51.4356375,7.4711599 51.4356610,7.4713450 51.4357210,7.4718180 51.4359665,7.4727150 51.4363345,7.4736552 51.4366234,7.4738781 51.4366918,7.4743868 51.4369218,7.4746695 51.4370459,7.4747952 51.4370759,7.4749620 51.4369720,7.4750489 51.4369187,7.4754839 51.4371570,7.4758479 51.4373453,7.4760679 51.4374592,7.4761999 51.4374179,7.4764529 51.4374484,7.4766051 51.4373019,7.4773551 51.4374108,7.4774061 51.4373308,7.4774467 51.4372719,7.4779999 51.4374617,7.4781281 51.4375051,7.4781813 51.4375231,7.4782323 51.4375404,7.4785334 51.4376423,7.4786050 51.4376665,7.4786560 51.4376837,7.4787069 51.4377010,7.4788142 51.4377373,7.4793016 51.4378987,7.4799592 51.4381164,7.4801252 51.4381742,7.4802385 51.4382135,7.4806985 51.4383736,7.4810191 51.4384850,7.4811114 51.4385171,7.4817251 51.4387306,7.4817741 51.4387472,7.4819918 51.4388209,7.4820810 51.4388535,7.4823692 51.4389546,7.4825479 51.4390173,7.4830146 51.4391833,7.4835692 51.4393766,7.4838754 51.4395003,7.4841684 51.4396318,7.4844285 51.4397555,7.4844707 51.4397756,7.4843995 51.4398367,7.4843612 51.4398708,7.4843398 51.4398899,7.4843416 51.4398907,7.4839051 51.4402781,7.4838969 51.4402880,7.4838849 51.4402960,7.4838650 51.4403138,7.4838363 51.4403393,7.4837956 51.4403754,7.4837515 51.4404145,7.4834805 51.4406550,7.4834396 51.4406914,7.4835743 51.4408642,7.4836806 51.4409728,7.4838102 51.4411150,7.4841597 51.4414810,7.4843999 51.4416595,7.4843991 51.4416604,7.4842587 51.4418198,7.4842297 51.4418527,7.4846514 51.4423501,7.4849733 51.4427824,7.4850155 51.4428322,7.4852978 51.4431647,7.4853800 51.4432970,7.4855318 51.4435409,7.4856884 51.4437779,7.4857601 51.4438941,7.4859979 51.4444372,7.4860244 51.4444975,7.4862227 51.4449312,7.4863557 51.4452499,7.4865140 51.4456833,7.4865214 51.4456971,7.4868356 51.4457855,7.4868357 51.4457861,7.4869506 51.4457609,7.4870226 51.4459176,7.4870833 51.4460712,7.4871541 51.4462360,7.4871785 51.4463518,7.4872662 51.4466595,7.4872961 51.4467689,7.4873055 51.4468609,7.4874197 51.4471690,7.4874497 51.4473027,7.4874862 51.4474133,7.4875266 51.4475353,7.4875916 51.4476614,7.4876016 51.4476806,7.4876522 51.4477874,7.4876800 51.4478926,7.4877024 51.4479499,7.4877733 51.4480599,7.4878185 51.4481256,7.4879904 51.4483407,7.4879979 51.4483555,7.4880237 51.4484070,7.4880964 51.4485564,7.4881672 51.4487158,7.4882217 51.4488455,7.4882610 51.4489980,7.4882755 51.4490864,7.4882819 51.4492525,7.4882921 51.4493563,7.4882970 51.4494572,7.4882961 51.4495098,7.4882802 51.4496272,7.4882699 51.4496998,7.4882598 51.4497521,7.4882432 51.4498081,7.4881995 51.4499521,7.4881257 51.4501351,7.4880992 51.4501844,7.4880494 51.4502564,7.4878569 51.4505341,7.4878150 51.4505927,7.4876793 51.4507690,7.4875371 51.4510444,7.4874798 51.4511840,7.4873884 51.4513038,7.4873129 51.4514177,7.4872113 51.4515611,7.4870638 51.4517846,7.4868586 51.4520494,7.4867513 51.4521827,7.4866392 51.4522987,7.4865526 51.4523760,7.4864672 51.4524464,7.4863766 51.4525138,7.4863278 51.4525466,7.4861632 51.4526449,7.4859405 51.4527687,7.4857038 51.4529059,7.4856738 51.4529216,7.4855102 51.4530080,7.4851246 51.4532224,7.4848506 51.4533889,7.4844386 51.4536527,7.4842389 51.4538215,7.4839456 51.4539621,7.4837445 51.4541506,7.4835032 51.4543969,7.4835081 51.4543881,7.4835016 51.4543963,7.4833641 51.4546226,7.4832855 51.4547964,7.4831784 51.4550992,7.4831502 51.4551995,7.4831228 51.4552979,7.4831069 51.4553655,7.4831906 51.4553600,7.4831872 51.4553819,7.4831878 51.4553818,7.4831455 51.4555901,7.4831363 51.4556972,7.4831407 51.4558029,7.4831628 51.4559470,7.4831956 51.4561609,7.4832285 51.4563754,7.4832732 51.4565891,7.4833272 51.4568019,7.4833855 51.4570136,7.4835566 51.4575772,7.4836680 51.4579221,7.4836877 51.4579811,7.4837403 51.4581546,7.4837394 51.4581554,7.4840644 51.4591729,7.4840674 51.4592826,7.4841478 51.4595511,7.4841679 51.4596181,7.4842280 51.4598188,7.4842937 51.4600905,7.4843491 51.4602530,7.4843758 51.4604301,7.4843885 51.4606093,7.4843896 51.4607933,7.4843701 51.4609669,7.4843496 51.4611494,7.4843087 51.4613205,7.4842575 51.4614973,7.4841994 51.4616735,7.4841164 51.4618767,7.4840861 51.4619094,7.4840421 51.4619356,7.4840080 51.4619452,7.4839889 51.4619536,7.4839510 51.4619586,7.4839297 51.4619608,7.4838345 51.4619555,7.4838279 51.4620091,7.4838194 51.4620780,7.4838661 51.4620826,7.4839119 51.4620844,7.4839688 51.4622018,7.4839254 51.4624321,7.4838571 51.4626715,7.4837731 51.4629658,7.4836329 51.4632494,7.4835580 51.4634007,7.4834370 51.4636575,7.4833520 51.4638400,7.4834076 51.4638500,7.4834460 51.4638602,7.4834370 51.4638805,7.4833874 51.4639959,7.4833032 51.4641843,7.4831756 51.4644605,7.4829229 51.4650990,7.4828404 51.4653073,7.4827808 51.4654580,7.4826952 51.4656224,7.4826765 51.4656584,7.4826535 51.4657022,7.4825839 51.4658362,7.4825727 51.4658578,7.4825558 51.4658750,7.4824887 51.4659373,7.4823838 51.4659413,7.4823808 51.4659592,7.4823779 51.4659771,7.4823750 51.4659944,7.4823404 51.4659873,7.4823202 51.4660270,7.4823212 51.4660289,7.4823161 51.4660313,7.4822879 51.4660903,7.4822497 51.4661653,7.4823283 51.4663278,7.4823232 51.4663508,7.4822943 51.4664318,7.4820581 51.4668545,7.4819104 51.4671379,7.4817780 51.4673469,7.4817376 51.4674201,7.4816714 51.4676729,7.4816750 51.4676728,7.4816672 51.4677275,7.4816632 51.4677554,7.4816624 51.4677749,7.4816615 51.4677862,7.4815721 51.4680341,7.4815067 51.4682447,7.4815042 51.4682458,7.4814453 51.4685286,7.4813939 51.4687752,7.4813452 51.4690087,7.4812960 51.4691277,7.4811969 51.4693674,7.4810621 51.4697545,7.4810581 51.4697659,7.4810359 51.4698577,7.4810272 51.4698934,7.4809657 51.4701474,7.4804303 51.4700361,7.4804510 51.4702128,7.4803904 51.4701998,7.4803773 51.4702175,7.4801658 51.4701818,7.4801101 51.4701695,7.4799345 51.4701309,7.4793844 51.4700194,7.4792028 51.4699806,7.4790798 51.4699554,7.4788094 51.4698999,7.4783707 51.4698097,7.4780579 51.4697456,7.4780314 51.4697397,7.4780141 51.4697760,7.4779686 51.4699402,7.4778596 51.4703262,7.4776914 51.4709230,7.4776868 51.4709391,7.4775857 51.4712972,7.4773913 51.4719255,7.4773747 51.4719792,7.4773701 51.4719936,7.4772005 51.4725418,7.4768736 51.4736063,7.4768437 51.4736785,7.4763751 51.4748123,7.4757138 51.4746885,7.4753794 51.4746201,7.4751370 51.4745707,7.4752388 51.4743408,7.4747540 51.4742513,7.4746633 51.4742344,7.4745115 51.4742120,7.4739386 51.4741327,7.4733956 51.4740570,7.4732581 51.4740378,7.4720103 51.4739147,7.4718765 51.4740975,7.4715120 51.4746324,7.4714755 51.4746867,7.4714735 51.4746864,7.4714738 51.4746868,7.4711986 51.4750893,7.4709619 51.4754354,7.4707116 51.4757909,7.4707062 51.4758037,7.4706922 51.4758195,7.4706721 51.4758653,7.4706545 51.4759253,7.4706061 51.4760394,7.4704835 51.4767571,7.4704399 51.4770714,7.4705173 51.4772307,7.4706887 51.4774773,7.4708487 51.4776638,7.4709638 51.4779570,7.4709967 51.4780430,7.4710085 51.4780670,7.4710126 51.4780848,7.4710216 51.4781050,7.4714851 51.4780712,7.4724384 51.4780278,7.4727155 51.4780151,7.4728359 51.4780108,7.4730891 51.4780010,7.4731123 51.4780001,7.4731250 51.4782198,7.4734781 51.4782058,7.4737829 51.4781905,7.4738257 51.4785377,7.4738576 51.4787968,7.4746590 51.4787655,7.4750812 51.4787489,7.4756360 51.4787113,7.4756601 51.4787096,7.4757240 51.4787053,7.4759211 51.4786919,7.4759720 51.4786100,7.4759762 51.4786642,7.4759785 51.4786880,7.4759674 51.4788182,7.4759276 51.4791324,7.4758862 51.4794257,7.4758396 51.4797468,7.4758407 51.4797743,7.4758423 51.4798186,7.4758437 51.4798505,7.4758734 51.4798774,7.4759433 51.4798877,7.4760015 51.4798861,7.4762719 51.4798787,7.4772510 51.4798208,7.4788844 51.4797242,7.4789018 51.4797232,7.4790116 51.4797167,7.4791874 51.4797063,7.4792803 51.4797007,7.4792802 51.4797009,7.4790909 51.4798790,7.4790510 51.4799157,7.4790115 51.4799008,7.4789667 51.4798976,7.4789213 51.4799023,7.4786192 51.4803075,7.4781715 51.4807576,7.4779781 51.4809181,7.4774884 51.4813097,7.4768872 51.4815809,7.4762844 51.4818502,7.4759343 51.4819957,7.4759130 51.4820041,7.4757057 51.4821041,7.4752792 51.4822896,7.4750242 51.4823961,7.4741992 51.4826978,7.4740499 51.4827529,7.4738166 51.4828396,7.4736609 51.4828962,7.4733388 51.4829938,7.4731320 51.4830404,7.4731310 51.4830413,7.4730538 51.4830587,7.4729427 51.4830875,7.4723962 51.4833020,7.4714713 51.4836709,7.4705491 51.4840453,7.4701458 51.4842137,7.4700493 51.4842520,7.4700384 51.4842705,7.4700338 51.4842872,7.4700339 51.4843044,7.4700383 51.4843204,7.4700490 51.4843384,7.4700630 51.4843529,7.4700759 51.4843628,7.4701084 51.4843788,7.4699583 51.4844587,7.4698815 51.4844997,7.4698099 51.4844832,7.4696416 51.4844549,7.4696170 51.4844517,7.4695865 51.4844476,7.4695703 51.4844482,7.4695841 51.4844492,7.4694321 51.4844807,7.4692088 51.4845269,7.4691960 51.4845325,7.4688408 51.4846859,7.4685367 51.4848937,7.4685351 51.4848923,7.4685348 51.4848934,7.4683752 51.4850702,7.4683223 51.4851371,7.4681208 51.4853912,7.4679856 51.4855644,7.4679631 51.4857248,7.4678668 51.4858919,7.4677748 51.4860014,7.4676857 51.4861074,7.4676586 51.4861437,7.4675993 51.4862235,7.4675878 51.4862390,7.4675795 51.4862503,7.4675243 51.4863247,7.4675111 51.4863421,7.4674784 51.4863850,7.4674495 51.4864279,7.4672628 51.4867046,7.4671119 51.4870084,7.4669418 51.4873081,7.4668037 51.4877362,7.4667558 51.4881393,7.4667403 51.4885667,7.4667492 51.4886545,7.4668412 51.4895624,7.4668453 51.4896027,7.4668519 51.4896672,7.4668558 51.4897072,7.4672110 51.4898526,7.4672114 51.4898136,7.4673271 51.4898603,7.4673273 51.4898613,7.4673788 51.4898821,7.4678087 51.4900553,7.4679081 51.4900953,7.4679151 51.4901013,7.4679152 51.4901368,7.4679156 51.4901371,7.4679175 51.4901378,7.4679175 51.4901373,7.4679189 51.4901383,7.4679161 51.4901763,7.4680338 51.4902159,7.4682340 51.4902867,7.4684702 51.4903502,7.4685510 51.4903635,7.4687097 51.4903898,7.4689370 51.4904076,7.4692181 51.4904021,7.4694800 51.4903782,7.4704705 51.4902199,7.4706232 51.4902075,7.4706502 51.4901833,7.4707260 51.4901832,7.4710820 51.4901341,7.4712261 51.4901142,7.4712725 51.4901447,7.4722258 51.4901605,7.4724299 51.4901277,7.4724788 51.4901198,7.4727361 51.4901284,7.4741294 51.4901717,7.4744234 51.4901909,7.4746915 51.4902577,7.4748617 51.4903329,7.4753769 51.4906119,7.4756200 51.4907246,7.4758872 51.4907947,7.4760032 51.4908087,7.4761781 51.4908299,7.4766323 51.4908637,7.4767583 51.4908908,7.4768057 51.4909151,7.4768315 51.4909467,7.4768682 51.4909934,7.4769316 51.4910628,7.4770170 51.4913103,7.4770169 51.4913112,7.4770071 51.4913665,7.4770446 51.4914139,7.4770922 51.4914459,7.4771158 51.4914751,7.4771539 51.4914993,7.4772077 51.4915195,7.4772158 51.4915209,7.4772896 51.4915339,7.4774090 51.4915486,7.4775455 51.4915408,7.4776403 51.4915314,7.4776503 51.4915304,7.4777538 51.4915174,7.4777997 51.4915363,7.4778285 51.4915213,7.4778613 51.4915080,7.4784607 51.4914357,7.4785885 51.4914220,7.4787368 51.4914095,7.4788060 51.4914037,7.4791756 51.4913590,7.4794476 51.4913262,7.4796269 51.4913063,7.4797350 51.4912974,7.4798425 51.4912967,7.4798881 51.4913011,7.4799306 51.4913118,7.4799694 51.4913138,7.4799994 51.4913233,7.4800417 51.4913368,7.4802533 51.4914217,7.4803889 51.4914696,7.4804489 51.4914943,7.4805137 51.4915136,7.4806644 51.4915528,7.4807466 51.4915809,7.4809161 51.4916641,7.4811118 51.4917554,7.4811782 51.4917886,7.4812427 51.4918314,7.4812978 51.4918629,7.4813777 51.4918986,7.4815213 51.4919734,7.4816360 51.4920334,7.4816983 51.4920657,7.4817277 51.4920846,7.4818580 51.4921696,7.4819330 51.4922266,7.4820043 51.4922779,7.4820188 51.4922906,7.4821110 51.4923707,7.4822009 51.4924291,7.4822650 51.4924707,7.4823338 51.4925079,7.4824069 51.4925473,7.4824941 51.4925848,7.4826800 51.4926796,7.4830859 51.4928695,7.4833324 51.4929881,7.4835820 51.4931082,7.4836685 51.4931497,7.4838653 51.4932440,7.4839921 51.4932918,7.4842676 51.4933936,7.4843234 51.4934143,7.4843767 51.4934457,7.4843992 51.4934590,7.4844640 51.4934812,7.4845642 51.4935031,7.4847493 51.4935508,7.4847716 51.4935217,7.4849156 51.4935390,7.4851737 51.4935572,7.4853169 51.4935322,7.4854985 51.4934617,7.4856839 51.4933966,7.4859080 51.4933178,7.4861383 51.4932229,7.4863643 51.4930291,7.4865648 51.4928103,7.4866977 51.4926646,7.4868461 51.4924998,7.4870913 51.4922275,7.4872094 51.4921265,7.4873266 51.4920879,7.4888270 51.4916755,7.4896329 51.4914356,7.4897145 51.4914094,7.4897738 51.4913905,7.4897636 51.4913819,7.4897326 51.4913555,7.4896941 51.4913227,7.4895903 51.4912341,7.4897374 51.4912250,7.4898688 51.4911787,7.4902492 51.4911604,7.4902858 51.4911904,7.4903300 51.4912266,7.4904222 51.4913586,7.4904726 51.4915294,7.4904577 51.4915577,7.4905096 51.4916347,7.4906241 51.4918043,7.4909760 51.4921873,7.4910326 51.4922568,7.4910724 51.4923058,7.4909818 51.4923344,7.4916975 51.4932033,7.4917873 51.4931748,7.4919196 51.4933322,7.4919648 51.4933289,7.4919978 51.4933739,7.4920084 51.4933886,7.4920713 51.4934778,7.4921164 51.4935310,7.4921374 51.4935560,7.4921669 51.4935904,7.4921852 51.4936123,7.4922384 51.4936813,7.4922954 51.4937553,7.4923103 51.4937746,7.4923389 51.4938117,7.4924937 51.4940235,7.4925181 51.4940569,7.4926013 51.4941709,7.4926208 51.4941936,7.4927056 51.4942926,7.4928139 51.4944314,7.4928371 51.4944614,7.4929005 51.4945443,7.4929400 51.4945958,7.4929742 51.4946404,7.4929846 51.4946561,7.4931018 51.4948328,7.4931119 51.4948481,7.4931348 51.4948884,7.4932470 51.4949259,7.4931921 51.4949847,7.4932227 51.4950472,7.4931908 51.4950722,7.4931981 51.4951027,7.4932345 51.4951698,7.4932804 51.4952515,7.4933001 51.4952722,7.4933467 51.4953277,7.4933612 51.4953651,7.4933888 51.4954347,7.4933959 51.4954527,7.4934213 51.4955168,7.4935598 51.4958582,7.4935661 51.4958739,7.4936274 51.4962001,7.4936583 51.4964185,7.4936671 51.4966601,7.4936736 51.4967233,7.4936766 51.4967532,7.4936920 51.4969063,7.4936980 51.4969504,7.4936971 51.4969655,7.4936932 51.4970626,7.4936878 51.4971647,7.4936719 51.4972647,7.4936404 51.4974421,7.4936037 51.4976492,7.4935721 51.4978280,7.4935628 51.4978802,7.4935543 51.4979283,7.4935294 51.4980686,7.4934676 51.4984178,7.4933497 51.4990034,7.4932648 51.4994211,7.4932161 51.4997192,7.4932083 51.4997771,7.4932433 51.4997672,7.4935547 51.4996796,7.4937335 51.4996433,7.4939106 51.4996041,7.4939750 51.4995893,7.4943125 51.4995103,7.4944035 51.4994882,7.4946486 51.4994291,7.4949842 51.4993477,7.4950869 51.4993230,7.4951449 51.4993090,7.4953506 51.4992593,7.4954769 51.4992289,7.4957332 51.4991675,7.4962141 51.4990521,7.4962127 51.4990031,7.4962127 51.4990024,7.4962534 51.4989927,7.4962898 51.4989840,7.4963924 51.4989594,7.4963975 51.4989589,7.4969414 51.4988289,7.4969642 51.4988235,7.4969780 51.4988193,7.4969815 51.4988177,7.4970845 51.4987939,7.4971922 51.4987715,7.4973928 51.4987298,7.4975370 51.4986990,7.4975517 51.4986965,7.4981243 51.4985744,7.4983131 51.4985333,7.4988180 51.4984231,7.4992853 51.4983244,7.4998215 51.4982110,7.4998347 51.4982073,7.4998974 51.4981945,7.5000445 51.4981646,7.5000644 51.4981595,7.5000696 51.4981588,7.5005314 51.4980602,7.5009036 51.4979817,7.5012955 51.4978991,7.5014132 51.4978742,7.5014341 51.4978698,7.5015439 51.4978467,7.5019206 51.4977672,7.5019612 51.4977585,7.5019699 51.4977558,7.5020707 51.4977334,7.5021270 51.4977224,7.5025575 51.4976294,7.5030656 51.4974418,7.5043189 51.4967620,7.5048931 51.4966000,7.5053343 51.4965451,7.5055672 51.4965336,7.5055692 51.4965567,7.5056032 51.4969589,7.5056099 51.4970353,7.5056272 51.4972239,7.5056740 51.4977325,7.5056950 51.4979610,7.5057561 51.4986279,7.5057612 51.4986855,7.5058029 51.4991434,7.5058050 51.4991761,7.5058068 51.4992066,7.5058272 51.4994176,7.5058439 51.4994854,7.5058791 51.4995704,7.5059231 51.4996536,7.5060218 51.4998137,7.5060589 51.4998737,7.5061099 51.4999568,7.5061558 51.5000393,7.5061939 51.5001226,7.5062212 51.5002055,7.5062306 51.5002704,7.5062257 51.5003703,7.5061949 51.5004688,7.5061618 51.5005316,7.5061063 51.5006090,7.5060940 51.5006228,7.5060566 51.5006652,7.5060398 51.5006841,7.5059860 51.5007375,7.5059648 51.5007587,7.5058864 51.5008324,7.5056938 51.5010099,7.5056826 51.5010180,7.5056669 51.5010247,7.5056490 51.5010288,7.5056318 51.5010296,7.5056130 51.5010277,7.5055957 51.5010227,7.5055913 51.5010210,7.5055940 51.5010251,7.5057087 51.5010688,7.5058610 51.5011310,7.5058636 51.5011286,7.5059880 51.5011775,7.5059895 51.5011781,7.5060851 51.5010777,7.5061353 51.5010231,7.5064411 51.5007182,7.5066714 51.5005369,7.5067496 51.5004814,7.5067836 51.5004573,7.5069142 51.5003663,7.5073232 51.5001235,7.5077151 51.4999477,7.5078359 51.4999040,7.5079823 51.4998511,7.5082892 51.4997546,7.5083117 51.4997475,7.5086670 51.4996648,7.5087520 51.4996450,7.5096305 51.4994994,7.5101989 51.4994052,7.5106388 51.4993395,7.5109575 51.4992920,7.5110392 51.4992791,7.5114186 51.4992194,7.5122519 51.4990936,7.5127212 51.4990227,7.5133735 51.4989242,7.5132790 51.4986254,7.5131920 51.4984227,7.5131755 51.4983841,7.5130866 51.4983515,7.5131575 51.4983399,7.5131866 51.4983349,7.5133015 51.4983153,7.5132840 51.4983351,7.5132472 51.4983729,7.5132642 51.4984125,7.5133508 51.4986137,7.5135163 51.4989035,7.5147541 51.4986994,7.5163894 51.4984111,7.5164198 51.4984058,7.5172614 51.4982573,7.5178792 51.4981461,7.5179223 51.4981415,7.5181558 51.4980931,7.5193880 51.4978486,7.5198580 51.4977116,7.5199852 51.4976453,7.5199116 51.4974080,7.5199062 51.4973901,7.5199211 51.4973883,7.5199214 51.4973884,7.5198684 51.4972147,7.5197969 51.4969807,7.5197921 51.4969650,7.5198094 51.4969664,7.5197042 51.4966258,7.5196579 51.4964758,7.5196015 51.4962903,7.5193996 51.4956389,7.5193021 51.4953244,7.5192854 51.4952690,7.5192617 51.4951921,7.5191992 51.4949894,7.5191052 51.4946851,7.5190282 51.4944290,7.5189873 51.4942933,7.5189634 51.4942142,7.5189519 51.4941760,7.5189477 51.4941620,7.5189320 51.4941099,7.5189079 51.4940297,7.5188971 51.4939962,7.5188697 51.4939104,7.5188527 51.4938673,7.5188228 51.4937636,7.5188944 51.4937331,7.5190198 51.4937470,7.5191161 51.4937577,7.5192368 51.4937766,7.5193374 51.4937924,7.5194188 51.4938052,7.5195476 51.4938254,7.5195461 51.4938242,7.5195451 51.4938042,7.5195737 51.4937368,7.5195806 51.4937208,7.5195690 51.4936462,7.5195694 51.4936463,7.5194560 51.4928924,7.5194521 51.4928668,7.5194434 51.4928553,7.5192622 51.4927363,7.5190928 51.4927301,7.5190750 51.4926871,7.5190503 51.4926352,7.5193480 51.4926954,7.5200387 51.4928352,7.5200729 51.4928420,7.5199965 51.4927818,7.5197969 51.4926247,7.5194964 51.4923997,7.5193853 51.4920147,7.5192107 51.4919073,7.5191153 51.4917460,7.5189209 51.4913550,7.5188829 51.4912787,7.5189001 51.4912479,7.5191694 51.4907665,7.5195911 51.4900123,7.5196491 51.4900492,7.5201398 51.4903617,7.5204055 51.4896086,7.5204482 51.4895278,7.5205009 51.4894433,7.5205669 51.4893611,7.5206196 51.4893045,7.5207212 51.4891955,7.5207901 51.4891340,7.5208455 51.4890923,7.5209129 51.4890532,7.5209819 51.4890230,7.5210584 51.4889972,7.5211394 51.4889784,7.5212797 51.4889585,7.5221544 51.4888347,7.5225661 51.4887894,7.5230313 51.4887289,7.5234049 51.4886812,7.5235281 51.4886655,7.5236846 51.4886546,7.5238424 51.4886622,7.5239951 51.4886873,7.5240284 51.4886970,7.5241379 51.4887289,7.5250717 51.4890879,7.5256276 51.4893036,7.5260097 51.4894471,7.5260858 51.4894758,7.5262759 51.4895472,7.5271656 51.4898838,7.5271754 51.4898874,7.5271829 51.4898681,7.5277031 51.4885384,7.5277870 51.4883237,7.5279164 51.4879921,7.5281191 51.4874734,7.5281517 51.4873901,7.5279504 51.4873520,7.5275092 51.4872438,7.5272183 51.4871653,7.5270202 51.4870780,7.5268822 51.4869737,7.5268778 51.4869750,7.5266438 51.4867553,7.5266096 51.4867231,7.5266508 51.4867260,7.5266756 51.4867296,7.5267070 51.4867116,7.5268151 51.4867356,7.5269132 51.4868253,7.5270143 51.4869098,7.5272469 51.4870576,7.5275411 51.4871569,7.5281466 51.4872634,7.5281473 51.4872631,7.5281430 51.4871222,7.5281365 51.4869190,7.5281353 51.4868758,7.5280984 51.4868693,7.5280637 51.4868631,7.5280813 51.4868378,7.5281315 51.4867601,7.5281218 51.4864464,7.5281199 51.4863862,7.5281120 51.4861391,7.5281063 51.4859533,7.5281025 51.4858322,7.5279913 51.4858163,7.5280991 51.4857237,7.5280896 51.4854182,7.5280885 51.4853821,7.5279457 51.4853471,7.5279106 51.4853816,7.5274902 51.4853753,7.5273153 51.4853705,7.5272397 51.4855176,7.5268018 51.4853247,7.5268004 51.4853312,7.5264154 51.4852582,7.5263710 51.4852495,7.5262184 51.4852187,7.5262190 51.4852136,7.5260420 51.4851535,7.5259772 51.4851299,7.5259002 51.4851032,7.5258453 51.4850844,7.5258131 51.4850760,7.5256451 51.4850319,7.5255548 51.4850081,7.5255505 51.4849165,7.5255458 51.4848134,7.5255447 51.4848005,7.5255439 51.4847671,7.5255216 51.4847611,7.5252566 51.4847130,7.5252389 51.4847098,7.5252260 51.4847373,7.5249723 51.4846895,7.5248283 51.4846625,7.5247934 51.4846559,7.5246722 51.4848809,7.5245296 51.4848795,7.5242813 51.4848708,7.5241543 51.4848616,7.5239793 51.4848489,7.5239792 51.4848351,7.5236725 51.4848095,7.5234052 51.4847736,7.5232605 51.4847522,7.5230185 51.4847354,7.5230125 51.4847441,7.5226487 51.4847388,7.5226336 51.4844393,7.5226316 51.4844394,7.5225857 51.4844412,7.5225032 51.4844449,7.5224535 51.4844448,7.5223032 51.4844430,7.5222769 51.4844497,7.5221302 51.4844792,7.5215263 51.4846011,7.5214904 51.4846083,7.5213500 51.4846366,7.5212638 51.4846541,7.5210962 51.4846878,7.5205545 51.4847971,7.5204184 51.4848246,7.5203650 51.4848355,7.5203975 51.4849267,7.5204319 51.4850229,7.5204972 51.4852061,7.5205376 51.4853192,7.5206275 51.4855714,7.5206934 51.4856897,7.5205096 51.4857187,7.5204569 51.4857271,7.5202706 51.4857565,7.5201046 51.4854092,7.5200784 51.4853713,7.5200179 51.4852833,7.5199849 51.4852261,7.5199664 51.4851915,7.5199594 51.4851787,7.5198990 51.4850665,7.5198324 51.4849428,7.5198051 51.4849483,7.5190758 51.4850954,7.5187486 51.4851616,7.5186928 51.4851729,7.5186962 51.4851838,7.5188901 51.4856054,7.5190438 51.4859394,7.5185334 51.4860236,7.5184727 51.4858867,7.5182414 51.4859421,7.5177045 51.4860433,7.5176827 51.4859151,7.5176143 51.4855105,7.5176084 51.4855128,7.5176017 51.4854762,7.5175219 51.4853995,7.5174498 51.4853304,7.5174455 51.4853170,7.5174465 51.4853166,7.5175551 51.4852701,7.5176330 51.4852381,7.5178408 51.4851562,7.5181430 51.4850331,7.5184358 51.4849139,7.5181648 51.4841755,7.5180745 51.4839013,7.5181081 51.4838960,7.5182914 51.4838665,7.5184679 51.4838421,7.5185834 51.4838281,7.5193908 51.4837295,7.5194609 51.4837221,7.5198316 51.4836907,7.5200231 51.4836744,7.5200831 51.4836693,7.5203175 51.4836495,7.5203548 51.4836463,7.5204457 51.4836386,7.5206136 51.4836263,7.5207221 51.4836220,7.5208394 51.4836175,7.5208415 51.4836167,7.5210656 51.4836063,7.5212470 51.4835978,7.5213070 51.4835565,7.5214774 51.4835581,7.5215665 51.4834780,7.5216738 51.4833815,7.5217640 51.4833005,7.5216393 51.4832495,7.5216054 51.4832123,7.5215832 51.4831878,7.5215750 51.4831788,7.5216546 51.4831476,7.5217547 51.4830791,7.5218432 51.4830226,7.5219467 51.4829885,7.5221337 51.4829498,7.5222018 51.4824624,7.5222479 51.4821326,7.5222915 51.4818209,7.5225491 51.4818789,7.5225498 51.4818784,7.5225606 51.4817878,7.5225729 51.4817007,7.5224470 51.4816699,7.5223155 51.4816501,7.5223156 51.4816492,7.5223143 51.4816493,7.5221095 51.4816113,7.5219214 51.4815872,7.5216487 51.4815619,7.5215273 51.4815477,7.5214502 51.4815081,7.5214500 51.4815082,7.5212372 51.4815049,7.5210223 51.4814945,7.5208396 51.4814936,7.5208226 51.4814936,7.5206679 51.4814928,7.5205030 51.4814961,7.5203253 51.4815045,7.5199964 51.4815321,7.5196790 51.4815651,7.5193196 51.4816025,7.5189542 51.4816390,7.5187772 51.4816652,7.5185324 51.4816944,7.5183419 51.4817186,7.5180961 51.4817501,7.5175314 51.4818251,7.5173119 51.4818542,7.5170765 51.4818868,7.5168006 51.4819251,7.5165440 51.4819606,7.5163590 51.4819830,7.5160476 51.4820318,7.5160380 51.4820662,7.5156438 51.4821205,7.5152005 51.4821537,7.5152044 51.4821245,7.5152758 51.4818496,7.5153174 51.4816730,7.5153446 51.4815574,7.5154858 51.4812763,7.5159495 51.4813950,7.5162841 51.4814688,7.5163406 51.4812743,7.5165270 51.4806321,7.5165989 51.4803845,7.5166246 51.4802962,7.5168311 51.4802578,7.5170803 51.4802114,7.5171386 51.4801699,7.5173131 51.4800458,7.5174645 51.4799380,7.5175071 51.4799283,7.5179116 51.4798360,7.5178952 51.4797587,7.5178921 51.4797445,7.5178814 51.4796940,7.5179804 51.4792897,7.5180069 51.4791812,7.5180121 51.4791606,7.5180223 51.4791190,7.5180289 51.4790927,7.5179602 51.4791121,7.5178959 51.4791232,7.5178341 51.4791337,7.5177397 51.4791447,7.5176682 51.4791530,7.5176115 51.4790472,7.5172898 51.4785575,7.5169793 51.4780848,7.5170239 51.4774223,7.5170634 51.4768347,7.5171322 51.4768808,7.5171626 51.4768099,7.5171618 51.4768094,7.5171899 51.4767464,7.5171649 51.4767337,7.5170672 51.4766846,7.5170669 51.4766841,7.5170079 51.4766512,7.5168960 51.4766167,7.5167884 51.4765839,7.5166270 51.4765626,7.5164385 51.4765527,7.5163718 51.4765083,7.5163680 51.4763280,7.5163659 51.4762332,7.5163614 51.4761902,7.5163525 51.4760900,7.5163454 51.4760103,7.5163325 51.4758324,7.5163241 51.4757706,7.5163295 51.4757254,7.5163295 51.4756900,7.5163584 51.4756036,7.5164036 51.4755381,7.5164150 51.4755244,7.5164306 51.4755056,7.5165013 51.4754450,7.5166224 51.4753542,7.5166968 51.4752956,7.5168518 51.4751645,7.5169337 51.4750791,7.5170344 51.4749785,7.5172217 51.4747916,7.5173013 51.4747121,7.5174014 51.4746114,7.5174032 51.4746053,7.5173915 51.4745024,7.5173795 51.4744092,7.5173431 51.4741776,7.5174067 51.4740226,7.5174788 51.4738461,7.5175297 51.4737222,7.5175872 51.4735814,7.5176539 51.4734188,7.5177620 51.4732683,7.5181468 51.4730075,7.5182620 51.4729667,7.5184618 51.4729371,7.5185086 51.4729411,7.5185438 51.4729442,7.5188667 51.4729719,7.5189022 51.4729750,7.5189202 51.4729766,7.5189646 51.4729804,7.5192584 51.4730057,7.5192915 51.4730071,7.5195990 51.4730194,7.5197992 51.4730209,7.5198678 51.4730214,7.5200331 51.4730193,7.5202324 51.4730168,7.5204691 51.4730101,7.5206591 51.4730053,7.5207114 51.4730040,7.5208697 51.4730002,7.5209327 51.4729987,7.5210540 51.4729959,7.5212015 51.4729787,7.5214778 51.4729464,7.5217637 51.4729285,7.5217623 51.4729280,7.5219490 51.4729313,7.5219628 51.4729073,7.5219763 51.4728872,7.5220214 51.4728200,7.5219955 51.4728226,7.5220135 51.4727824,7.5220464 51.4727086,7.5220661 51.4726644,7.5220703 51.4726551,7.5221886 51.4723901,7.5221947 51.4723761,7.5221980 51.4723761,7.5222988 51.4723750,7.5224426 51.4723736,7.5225441 51.4723733,7.5228108 51.4723870,7.5229663 51.4724024,7.5230995 51.4724157,7.5231646 51.4724114,7.5232440 51.4724060,7.5232719 51.4723563,7.5232861 51.4723311,7.5233512 51.4722173,7.5234268 51.4722394,7.5234505 51.4722009,7.5234834 51.4721474,7.5234855 51.4721470,7.5236090 51.4719467,7.5237740 51.4716551,7.5237733 51.4716527,7.5238309 51.4715572,7.5240024 51.4712629,7.5241356 51.4710348,7.5242139 51.4708967,7.5242174 51.4708944,7.5243985 51.4705837,7.5244142 51.4705569,7.5245376 51.4703489,7.5245361 51.4703472,7.5245389 51.4703423,7.5246171 51.4702077,7.5246200 51.4702064,7.5247703 51.4699467,7.5247696 51.4699435,7.5248011 51.4698890,7.5248081 51.4698772,7.5248279 51.4698431,7.5248481 51.4698085,7.5248502 51.4698086,7.5250091 51.4695359,7.5250497 51.4694674,7.5250814 51.4694100,7.5250650 51.4694000,7.5250667 51.4693999,7.5250053 51.4693581,7.5250087 51.4693514,7.5250107 51.4693515,7.5250329 51.4693063,7.5252844 51.4688007,7.5252877 51.4687873,7.5252941 51.4687744,7.5253086 51.4687450,7.5253222 51.4687169,7.5253196 51.4687144,7.5253724 51.4686166,7.5254252 51.4685127,7.5254263 51.4685136,7.5254742 51.4684173,7.5256253 51.4681444,7.5256709 51.4680620,7.5257388 51.4679589,7.5257612 51.4679249,7.5258159 51.4678419,7.5258870 51.4677338,7.5259370 51.4676579,7.5260255 51.4675234,7.5260564 51.4674766,7.5261876 51.4672649,7.5262092 51.4672300,7.5262365 51.4671742,7.5263072 51.4669911,7.5263263 51.4669416,7.5263805 51.4668015,7.5264303 51.4666725,7.5265044 51.4664808,7.5265623 51.4663309,7.5266059 51.4662182,7.5266795 51.4660067,7.5266933 51.4659650,7.5266912 51.4659651,7.5266923 51.4659619,7.5266971 51.4659473,7.5266996 51.4659471,7.5267605 51.4657652,7.5267947 51.4656689,7.5267994 51.4656552,7.5268197 51.4655976,7.5268499 51.4655116,7.5268821 51.4654197,7.5268934 51.4653879,7.5269579 51.4652043,7.5269702 51.4651689,7.5269972 51.4650918,7.5270332 51.4649894,7.5270650 51.4648990,7.5270721 51.4648773,7.5270957 51.4648043,7.5271405 51.4646988,7.5271559 51.4646621,7.5271718 51.4646246,7.5272178 51.4645348,7.5272596 51.4644533,7.5272674 51.4644384,7.5273465 51.4642952,7.5275178 51.4641430,7.5277374 51.4639561,7.5277522 51.4639433,7.5278028 51.4639051,7.5278410 51.4638855,7.5279714 51.4638187,7.5280583 51.4637675,7.5282503 51.4636584,7.5282859 51.4636382,7.5283244 51.4636163,7.5283232 51.4636155,7.5283544 51.4635878,7.5283545 51.4635869,7.5283557 51.4635871,7.5283801 51.4635733,7.5284063 51.4635570,7.5284228 51.4635467,7.5285109 51.4634918,7.5288139 51.4633029,7.5289959 51.4631372,7.5291336 51.4629908,7.5291507 51.4629729,7.5292001 51.4629150,7.5292647 51.4628398,7.5293025 51.4627954,7.5293878 51.4626957,7.5293982 51.4626837,7.5294603 51.4626108,7.5294689 51.4626010,7.5296235 51.4624140,7.5296996 51.4623221,7.5297385 51.4622750,7.5298775 51.4620818,7.5300177 51.4618869,7.5302517 51.4615794,7.5303752 51.4614171,7.5305430 51.4612430,7.5305882 51.4611961,7.5306218 51.4611613,7.5306519 51.4611300,7.5308345 51.4609482,7.5309432 51.4608400,7.5312799 51.4605082,7.5312736 51.4605093,7.5312741 51.4605081,7.5313254 51.4604581,7.5313536 51.4604314,7.5313697 51.4604159,7.5313743 51.4604115,7.5314251 51.4603645,7.5314279 51.4603640,7.5315097 51.4602846,7.5315253 51.4602695,7.5315398 51.4602554,7.5316550 51.4601435,7.5317210 51.4600807,7.5320429 51.4597627,7.5321071 51.4596992,7.5322063 51.4596035,7.5322303 51.4595805,7.5323960 51.4594207,7.5324096 51.4594077,7.5324984 51.4593221,7.5326386 51.4593150,7.5332826 51.4592992,7.5333717 51.4593022,7.5337276 51.4593147,7.5338287 51.4593184,7.5340929 51.4593285,7.5343909 51.4593398,7.5344169 51.4593409,7.5346043 51.4593437,7.5349404 51.4593642,7.5350143 51.4593679,7.5352167 51.4593779,7.5353212 51.4593831,7.5355679 51.4593953,7.5356553 51.4593996,7.5356883 51.4594011,7.5358878 51.4594061,7.5360073 51.4594095,7.5361191 51.4594124,7.5361181 51.4594324,7.5363624 51.4594527,7.5368490 51.4594863,7.5368803 51.4594810,7.5369104 51.4594758,7.5369725 51.4594654,7.5370007 51.4594606,7.5371433 51.4594594,7.5375155 51.4595230,7.5375505 51.4595290,7.5375445 51.4595376,7.5375890 51.4595504,7.5375950 51.4595420,7.5381402 51.4596929,7.5381373 51.4597025,7.5383111 51.4597533,7.5384946 51.4598074,7.5386770 51.4598613,7.5388572 51.4599147,7.5392356 51.4600270,7.5394618 51.4600942,7.5395004 51.4601058,7.5398284 51.4602032,7.5398859 51.4602183,7.5403573 51.4603431,7.5405188 51.4603886,7.5408035 51.4604688,7.5415572 51.4606812,7.5416021 51.4606939,7.5419333 51.4607982,7.5420721 51.4608486,7.5424010 51.4609728,7.5424479 51.4609906,7.5425897 51.4608481,7.5429024 51.4608839,7.5430894 51.4608499,7.5432184 51.4608269,7.5433924 51.4607969,7.5436773 51.4607779,7.5441843 51.4607469,7.5443882 51.4607739,7.5446872 51.4608159,7.5448891 51.4608228,7.5449231 51.4605790,7.5449461 51.4603960,7.5449981 51.4600030,7.5450871 51.4599380,7.5451741 51.4599080,7.5454341 51.4598200,7.5455561 51.4597741,7.5460930 51.4595871,7.5461810 51.4594801,7.5463039 51.4593321,7.5464979 51.4591681,7.5466259 51.4590881,7.5472398 51.4586982,7.5475348 51.4585112,7.5476043 51.4584968,7.5476048 51.4584662,7.5476767 51.4584202,7.5479117 51.4582712,7.5480197 51.4582252,7.5485156 51.4580233,7.5494345 51.4576493,7.5496415 51.4573403,7.5498514 51.4570314,7.5499584 51.4568754,7.5501184 51.4567074,7.5504113 51.4564054,7.5505523 51.4562605,7.5507013 51.4561275,7.5508089 51.4560053,7.5509186 51.4559242,7.5510552 51.4558645,7.5512814 51.4556815,7.5515742 51.4555095,7.5520781 51.4553036,7.5525648 51.4550790,7.5523968 51.4550345,7.5522991 51.4550096,7.5519671 51.4549186,7.5518211 51.4548766,7.5515742 51.4548076,7.5512292 51.4547046,7.5499304 51.4543267,7.5497165 51.4542697,7.5482807 51.4538807,7.5469999 51.4535368,7.5467399 51.4534688,7.5465459 51.4533958,7.5453301 51.4529458,7.5448011 51.4527487,7.5442973 51.4525609,7.5440843 51.4524649,7.5440093 51.4524349,7.5439163 51.4523929,7.5435344 51.4522249,7.5432964 51.4521179,7.5428195 51.4518549,7.5420706 51.4514390,7.5416327 51.4511950,7.5415137 51.4511000,7.5411927 51.4508630,7.5411297 51.4508171,7.5406548 51.4504511,7.5399349 51.4498982,7.5389291 51.4491272,7.5388551 51.4490742,7.5387111 51.4489593,7.5385701 51.4489253,7.5384921 51.4486773,7.5383972 51.4485283,7.5385451 51.4484443,7.5381012 51.4480634,7.5374703 51.4470105,7.5371824 51.4465295,7.5370494 51.4463156,7.5369824 51.4462016,7.5368544 51.4459916,7.5365854 51.4455566,7.5365205 51.4454497,7.5364905 51.4453737,7.5363485 51.4450647,7.5362285 51.4447977,7.5361975 51.4447097,7.5361025 51.4444278,7.5359815 51.4440688,7.5357846 51.4434819,7.5356746 51.4431309,7.5354546 51.4424210,7.5354056 51.4422610,7.5353336 51.4420280,7.5350737 51.4414181,7.5349427 51.4411131,7.5348887 51.4409832,7.5346647 51.4404562,7.5347237 51.4403112,7.5343668 51.4397473,7.5342738 51.4394383,7.5341578 51.4391294,7.5339319 51.4385534,7.5338289 51.4382895,7.5337219 51.4380455,7.5336809 51.4379545,7.5334619 51.4374505,7.5332940 51.4370656,7.5332120 51.4368746,7.5330930 51.4366186,7.5328720 51.4361497,7.5327211 51.4358297,7.5325091 51.4355168,7.5324631 51.4354438,7.5324331 51.4352878,7.5322451 51.4351088,7.5321521 51.4350168,7.5320882 51.4349748,7.5319802 51.4349138,7.5317952 51.4348068,7.5315152 51.4346659,7.5313773 51.4346129,7.5312213 51.4345479,7.5310113 51.4344829,7.5306714 51.4343839,7.5304764 51.4343689,7.5298385 51.4343189,7.5293416 51.4343339,7.5292636 51.4343379,7.5290946 51.4343609,7.5290106 51.4343719,7.5287826 51.4344029,7.5285627 51.4344369,7.5283197 51.4344909,7.5269939 51.4350168,7.5254071 51.4356237,7.5251942 51.4357037,7.5250642 51.4357527,7.5246903 51.4359557,7.5244053 51.4361117,7.5242093 51.4362187,7.5234944 51.4366076,7.5231375 51.4367764,7.5228295 51.4364056,7.5214397 51.4355207,7.5213368 51.4354827,7.5202089 51.4350548,7.5184432 51.4346618,7.5180952 51.4346278,7.5179943 51.4346198,7.5174903 51.4345708,7.5167194 51.4344979,7.5165645 51.4344829,7.5151877 51.4342269,7.5142200 51.4340880,7.5141342 51.4341104,7.5141473 51.4340690,7.5140259 51.4339979,7.5138619 51.4339529,7.5132800 51.4337539,7.5124181 51.4334610,7.5118592 51.4333230,7.5117302 51.4331930,7.5117022 51.4330640,7.5116622 51.4328880,7.5115352 51.4325491,7.5114952 51.4324151,7.5113483 51.4318961,7.5111853 51.4313162,7.5111513 51.4311942,7.5110453 51.4308093,7.5107804 51.4298294,7.5106614 51.4297254,7.5105524 51.4296304,7.5102464 51.4293524,7.5089386 51.4281386,7.5085267 51.4277576,7.5081627 51.4276776,7.5079548 51.4276316,7.5070629 51.4274406,7.5068959 51.4274066,7.5068699 51.4274010,7.5065920 51.4273416,7.5062880 51.4272617,7.5060381 51.4271927,7.5056181 51.4270857,7.5050782 51.4269527,7.5046692 51.4268817,7.5058103 51.4261351,7.5057202 51.4259692,7.5058577 51.4253796,7.5057649 51.4249491,7.5049661 51.4247943,7.5057681 51.4242440,7.5065640 51.4236491,7.5073089 51.4230581,7.5074249 51.4229661,7.5074129 51.4228861,7.5073479 51.4227412,7.5072229 51.4224622,7.5071939 51.4220662,7.5071499 51.4214513,7.5071989 51.4213793,7.5072949 51.4212383,7.5073869 51.4212263,7.5075119 51.4212073,7.5079008 51.4212343,7.5081108 51.4212493,7.5081708 51.4212003,7.5082458 51.4211353,7.5083157 51.4209754,7.5083857 51.4207804,7.5085367 51.4203644,7.5086807 51.4199335,7.5087887 51.4196095,7.5088727 51.4193575,7.5089947 51.4192396,7.5093456 51.4191476,7.5098704 51.4190775,7.5102985 51.4189076,7.5105574 51.4188196,7.5111715 51.4185717,7.5115796 51.4183895,7.5119278 51.4182672,7.5117323 51.4181989,7.5117648 51.4181410,7.5115562 51.4180812,7.5112253 51.4181382,7.5108239 51.4178415,7.5093590 51.4180987,7.5088336 51.4181877,7.5086998 51.4182128,7.5075669 51.4184069,7.5063117 51.4185474,7.5056554 51.4185622,7.5053521 51.4185315,7.5048532 51.4185614,7.5045972 51.4185677,7.5041367 51.4185568,7.5034800 51.4185526,7.5033585 51.4185526,7.5030835 51.4185486,7.5028866 51.4185446,7.5026746 51.4185216,7.5020407 51.4185066,7.5017877 51.4185026,7.5014568 51.4184346,7.5011488 51.4183696,7.5010528 51.4183466,7.4994561 51.4179917,7.4991241 51.4179197,7.4989632 51.4178697,7.4988832 51.4178467,7.4978473 51.4175307,7.4975424 51.4174347,7.4966135 51.4171148,7.4963336 51.4170188,7.4961066 51.4169118,7.4951457 51.4164659,7.4948638 51.4163359,7.4935680 51.4157339,7.4932290 51.4155769,7.4931172 51.4155255,7.4930001 51.4156569,7.4926721 51.4158939,7.4924741 51.4160349,7.4922212 51.4161149,7.4921292 51.4161649,7.4918572 51.4162599,7.4913873 51.4164279,7.4909674 51.4165578,7.4890686 51.4171408,7.4889177 51.4171948,7.4876138 51.4176367,7.4874639 51.4176787,7.4872309 51.4177397,7.4868810 51.4178277,7.4866940 51.4178777,7.4864320 51.4179577,7.4856771 51.4178427,7.4855332 51.4178317,7.4853422 51.4178127,7.4848993 51.4177667,7.4846623 51.4177397,7.4834915 51.4176067,7.4830925 51.4175987,7.4824506 51.4175987,7.4791471 51.4175917,7.4784194 51.4176279,7.4777613 51.4175917,7.4776333 51.4175947,7.4775354 51.4175947,7.4771724 51.4175917,7.4763745 51.4175917,7.4762475 51.4175917,7.4759826 51.4175757,7.4758616 51.4175687,7.4753907 51.4175417,7.4751337 51.4175267,7.4750167 51.4175187,7.4744638 51.4174887,7.4739439 51.4174577,7.4738649 51.4174537,7.4720722 51.4173547,7.4718462 51.4173437,7.4712833 51.4173087,7.4696275 51.4171907,7.4688436 51.4171337,7.4687127 51.4171217,7.4669229 51.4169468,7.4667890 51.4169318,7.4662840 51.4168938,7.4662040 51.4168898,7.4660701 51.4168778,7.4658451 51.4168628,7.4656321 51.4168478,7.4653112 51.4168628,7.4650972 51.4168628,7.4649972 51.4169928,7.4649205 51.4170991,7.4647483 51.4172097,7.4646453 51.4172747,7.4644913 51.4173927,7.4641433 51.4179346,7.4639014 51.4183086,7.4638814 51.4185336,7.4638454 51.4186176,7.4638124 51.4188005,7.4637464 51.4190525,7.4637214 51.4192465,7.4636964 51.4194415)))");
 
     // initialize randomness
     srand(time(NULL) + rand());  // NOLINT
@@ -567,10 +143,6 @@ int main(int argc, char **argv) {
         }
     }
 
-    char *buf = new char[1024 * 1024 * 100];
-
-    size_t len;
-
     std::string dangling;
 
     size_t NUM_THREADS = std::thread::hardware_concurrency();
@@ -578,18 +150,10 @@ int main(int argc, char **argv) {
     Sweeper sweeper({NUM_THREADS, prefix, intersects, contains, suffix, useBoxIds, useArea}, useCache,
                     cache, output);
 
-    size_t gid = 0;
-
     if (!useCache) {
         LOGTO(INFO, std::cerr) << "Parsing input geometries...";
 
         parseStdin(sweeper);
-
-        /*
-        while ((len = read(0, buf, 1024 * 1024 * 100)) > 0) {
-            parse(buf, len, dangling, &gid, sweeper);
-        }
-         */
 
         sweeper.flush();
     }
