+++ src/spatialjoin/WktParser.h
@@ -1,14 +1,13 @@
-//
-// Created by kalmbacj on 4/23/24.
-//
+// Copyright 2023, University of Freiburg,
+// Chair of Algorithms and Data Structures.
+// Authors: Johannes Kalmbach <kalmbach@informatik.uni-freiburg.de>
 
 #ifndef SPATIALJOIN_WKTPARSER_H
 #define SPATIALJOIN_WKTPARSER_H
 
+#include "Sweeper.h"
 
-class WktParser {
-
-};
+void parseStdin(sj::Sweeper& idx);
 
 
 #endif //SPATIALJOIN_WKTPARSER_H
